Fix a trivial compilation error

+++ core/src/KVLogStore.cpp
@@ -151,7 +151,7 @@ int32_t KVLogStore::append(int64_t key, const std::string& value) {
     // Update the index
 
     // min with 0, since data_pos can be small (or zero) initially
-    for(uint64_t i = std::min(data_pos - ngram_n, 0ULL);
+    for(uint64_t i = std::min(data_pos - ngram_n, static_cast<uint64_t>(0));
         i < data_pos + val.length() - ngram_n; i++)
     {
         std::string ngram;
